refractor home_page code
